Project refactoring
