+++ config/example_config.yaml
@@ -17,9 +17,8 @@ datasets:
   # CIFAR-10 dataset
   - name: "CIFAR10"
     type: "cifar10"
-    root_path: "./data/cifar10"
+    root_path: "/mnt/d/data/cifar10"
     split: "test"  # train or test
-
   # ImageNet dataset
   - name: "ImageNet"
     type: "imagenet"
@@ -27,35 +26,23 @@ datasets:
     split: "val"
     class_map_file: "/path/to/imagenet/class_map.json"  # Optional
 
-  # Custom JSON format dataset
-  - name: "CustomDataset"
-    type: "custom"
-    root_path: "/path/to/custom/dataset"
-    split: "test"
-    annotation_file: "/path/to/custom/dataset/test_annotations.json"
-    annotation_format: "json"  # json, csv, or txt
-    image_key: "image_path"
-    label_key: "category"
-  # Custom CSV format dataset
-  - name: "CSVDataset"
-    type: "custom"
-    root_path: "/path/to/csv/dataset"
-    annotation_file: "/path/to/csv/dataset/labels.csv"
-    annotation_format: "csv"
-    image_dir: "/path/to/csv/dataset/images"
-    image_key: "filename"
-    label_key: "class_name"
+  # SUN397 dataset  
+  - name: "SUN397"
+    type: "sun397"
+    root_path: "/mnt/d/data/sun397/cache"  # HuggingFace cache directory
+    split: "train"  # Will auto-split for evaluation
+    # max_samples: 1000  # Optional: limit samples for testing
 
   # CIFAR-100 dataset with fine labels (100 classes)
   - name: "CIFAR100_Fine"
     type: "cifar100"
-    root_path: "./data/cifar100"
+    root_path: "/mnt/d/data/cifar100"
     split: "test"  # train or test
     use_coarse_labels: false  # Use 100 fine-grained classes
 
   # CIFAR-100 dataset with coarse labels (20 superclasses)
   - name: "CIFAR100_Coarse"
     type: "cifar100"
-    root_path: "./data/cifar100"
+    root_path: "/mnt/d/data/cifar100"
     split: "test"  # train or test
     use_coarse_labels: true  # Use 20 coarse-grained classes
