+++ results/cifar100_evaluation/config_used.yaml
@@ -9,11 +9,11 @@ save_results: true
 datasets:
 - name: CIFAR100_Fine
   type: cifar100
-  root_path: ./data/cifar100
+  root_path: /mnt/d/data/cifar100
   split: test
   use_coarse_labels: false
 - name: CIFAR100_Coarse
   type: cifar100
-  root_path: ./data/cifar100
+  root_path: /mnt/d/data/cifar100
   split: test
   use_coarse_labels: true
