+++ results/cifar100_evaluation/summary_20250623_174312.txt
@@ -13,4 +13,5 @@ Dataset Results:
 ------------------------------------------------------------
 Dataset                        Accuracy   Top-5      Per-Class 
 ------------------------------------------------------------
-CIFAR10                        89.58      99.43      89.58     
+CIFAR100_Fine                  0.64       0.88       0.64      
+CIFAR100_Coarse                0.50       0.91       0.50
