--- results/cifar10_evaluation/results_20250618_185957.json
@@ -1,153 +1,0 @@
-{
-  "CIFAR10": {
-    "dataset_name": "CIFAR10Adapter",
-    "num_samples": 10000,
-    "num_classes": 10,
-    "accuracy": 0.895799994468689,
-    "top5_accuracy": 0.9943,
-    "per_class_accuracy": {
-      "airplane": 0.9139999747276306,
-      "automobile": 0.9399999976158142,
-      "bird": 0.878000020980835,
-      "cat": 0.8619999885559082,
-      "deer": 0.8159999847412109,
-      "dog": 0.8870000243186951,
-      "frog": 0.781000018119812,
-      "horse": 0.9729999899864197,
-      "ship": 0.9629999995231628,
-      "truck": 0.9440000057220459
-    },
-    "mean_per_class_accuracy": 0.895799994468689,
-    "confusion_matrix": [
-      [
-        914,
-        7,
-        39,
-        0,
-        0,
-        1,
-        0,
-        0,
-        32,
-        7
-      ],
-      [
-        2,
-        940,
-        0,
-        0,
-        0,
-        1,
-        0,
-        0,
-        2,
-        55
-      ],
-      [
-        11,
-        2,
-        878,
-        21,
-        34,
-        13,
-        5,
-        33,
-        3,
-        0
-      ],
-      [
-        2,
-        7,
-        15,
-        862,
-        12,
-        78,
-        8,
-        11,
-        5,
-        0
-      ],
-      [
-        3,
-        2,
-        18,
-        12,
-        816,
-        17,
-        7,
-        122,
-        3,
-        0
-      ],
-      [
-        0,
-        1,
-        16,
-        66,
-        14,
-        887,
-        1,
-        12,
-        3,
-        0
-      ],
-      [
-        1,
-        4,
-        52,
-        79,
-        27,
-        32,
-        781,
-        9,
-        15,
-        0
-      ],
-      [
-        4,
-        1,
-        4,
-        4,
-        7,
-        6,
-        0,
-        973,
-        1,
-        0
-      ],
-      [
-        22,
-        10,
-        1,
-        0,
-        0,
-        0,
-        0,
-        0,
-        963,
-        4
-      ],
-      [
-        0,
-        43,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        13,
-        944
-      ]
-    ],
-    "model_info": {
-      "model_name": "ViT-B/32",
-      "device": "cuda",
-      "use_ensemble": true,
-      "vision_width": 224,
-      "vision_layers": 12,
-      "embed_dim": 512
-    },
-    "evaluation_time": "2025-06-18T18:59:57.784743"
-  }
-}
