+++ results/cifar10_evaluation/results_20250623_174230.json
@@ -3,104 +3,104 @@
     "dataset_name": "CIFAR10Adapter",
     "num_samples": 10000,
     "num_classes": 10,
-    "accuracy": 0.895799994468689,
-    "top5_accuracy": 0.9943,
+    "accuracy": 0.8956999778747559,
+    "top5_accuracy": 0.9955,
     "per_class_accuracy": {
       "airplane": 0.9139999747276306,
-      "automobile": 0.9399999976158142,
-      "bird": 0.878000020980835,
-      "cat": 0.8619999885559082,
-      "deer": 0.8159999847412109,
-      "dog": 0.8870000243186951,
-      "frog": 0.781000018119812,
+      "automobile": 0.9440000057220459,
+      "bird": 0.8730000257492065,
+      "cat": 0.871999979019165,
+      "deer": 0.8149999976158142,
+      "dog": 0.8859999775886536,
+      "frog": 0.7870000004768372,
       "horse": 0.9729999899864197,
-      "ship": 0.9629999995231628,
-      "truck": 0.9440000057220459
+      "ship": 0.9549999833106995,
+      "truck": 0.9380000233650208
     },
-    "mean_per_class_accuracy": 0.895799994468689,
+    "mean_per_class_accuracy": 0.8957000970840454,
     "confusion_matrix": [
       [
         914,
-        7,
-        39,
-        0,
-        0,
-        1,
-        0,
-        0,
-        32,
-        7
-      ],
-      [
-        2,
-        940,
-        0,
-        0,
-        0,
-        1,
-        0,
-        0,
-        2,
-        55
-      ],
-      [
-        11,
-        2,
-        878,
-        21,
-        34,
-        13,
-        5,
-        33,
-        3,
-        0
-      ],
-      [
-        2,
-        7,
-        15,
-        862,
-        12,
-        78,
         8,
-        11,
-        5,
+        48,
+        0,
+        0,
+        1,
+        1,
+        0,
+        23,
+        5
+      ],
+      [
+        2,
+        944,
+        0,
+        0,
+        0,
+        1,
+        0,
+        0,
+        2,
+        51
+      ],
+      [
+        12,
+        1,
+        873,
+        24,
+        36,
+        15,
+        6,
+        31,
+        2,
+        0
+      ],
+      [
+        2,
+        6,
+        13,
+        872,
+        10,
+        74,
+        9,
+        10,
+        4,
         0
       ],
       [
         3,
-        2,
+        1,
         18,
-        12,
-        816,
-        17,
-        7,
-        122,
+        16,
+        815,
+        21,
+        6,
+        117,
         3,
         0
       ],
       [
         0,
         1,
-        16,
-        66,
-        14,
-        887,
-        1,
+        15,
+        72,
         12,
-        3,
+        886,
+        1,
+        11,
+        2,
         0
       ],
       [
         1,
         4,
-        52,
-        79,
-        27,
-        32,
-        781,
+        46,
+        89,
+        22,
+        29,
+        787,
         9,
-        15,
+        13,
         0
       ],
       [
@@ -109,35 +109,35 @@
         4,
         4,
         7,
-        6,
+        7,
         0,
         973,
-        1,
+        0,
         0
       ],
       [
-        22,
+        27,
         10,
+        2,
+        0,
+        1,
         1,
         0,
         0,
-        0,
-        0,
-        0,
-        963,
+        955,
         4
       ],
       [
-        0,
-        43,
-        0,
-        0,
+        2,
+        49,
         0,
         0,
         0,
         0,
-        13,
-        944
+        0,
+        0,
+        11,
+        938
       ]
     ],
     "model_info": {
@@ -148,6 +148,6 @@
       "vision_layers": 12,
       "embed_dim": 512
     },
-    "evaluation_time": "2025-06-18T18:52:10.261469"
+    "evaluation_time": "2025-06-23T17:42:30.616653"
   }
 }
