+++ results/cifar10_evaluation/summary_20250623_174230.txt
@@ -13,5 +13,4 @@ Dataset Results:
 ------------------------------------------------------------
 Dataset                        Accuracy   Top-5      Per-Class 
 ------------------------------------------------------------
-CIFAR100_Fine                  63.96      88.11      63.96     
-CIFAR100_Coarse                49.97      90.29      49.97     
+CIFAR10                        0.90       1.00       0.90
