+++ results/sun397_evaluation/summary_20250623_173435.txt
@@ -13,4 +13,4 @@ Dataset Results:
 ------------------------------------------------------------
 Dataset                        Accuracy   Top-5      Per-Class 
 ------------------------------------------------------------
-CIFAR10                        89.58      99.43      89.58     
+SUN397                         0.50       0.71       0.48
